development/src/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.7
langchain==0.1.0
langchain-community==0.0.8
langchain-ollama==0.1.0
langchain-huggingface==0.0.1
chromadb==0.4.18
pypdf==3.17.1
python-multipart==0.0.6
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0


# ===== .env =====
# Database Configuration
DB_USER=postgres
DB_PASSWORD=your_password_here
DB_HOST=localhost
DB_PORT=5432
DB_NAME=rag_chatbot

# JWT Configuration
JWT_SECRET_KEY=your-super-secret-jwt-key-here
JWT_ALGORITHM=HS256
JWT_ACCESS_TOKEN_EXPIRE_MINUTES=30

# Ollama Configuration
OLLAMA_MODEL=phi3:mini
OLLAMA_BASE_URL=http://localhost:11434

# Vector Database
CHROMA_PATH=./chroma_db

# File Storage
UPLOAD_DIR=./uploads